[FEAT] #5 : 도전과제 1 - 갤러리에서 받아온 이미지 Uri를 Glide를 이용해 화면에 표시하기
