+++ app/src/main/java/co/kr/sopt_seminar_30th/presentation/ui/home/more/MoreFragment.kt
@@ -1,60 +1,145 @@
 package co.kr.sopt_seminar_30th.presentation.ui.home.more
 
+import android.Manifest
+import android.app.Activity.RESULT_CANCELED
+import android.app.Activity.RESULT_OK
+import android.app.AlertDialog
+import android.content.Intent
+import android.content.pm.PackageManager
 import android.os.Bundle
-import androidx.fragment.app.Fragment
+import android.provider.MediaStore
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Toast
+import androidx.activity.result.ActivityResultLauncher
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.core.content.ContextCompat
+import androidx.fragment.app.activityViewModels
 import co.kr.sopt_seminar_30th.R
+import co.kr.sopt_seminar_30th.databinding.FragmentMoreBinding
+import co.kr.sopt_seminar_30th.presentation.ui.base.BaseFragment
+import co.kr.sopt_seminar_30th.presentation.viewmodel.HomeViewModel
 
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
+class MoreFragment : BaseFragment<FragmentMoreBinding>() {
+    override val TAG: String
+        get() = MoreFragment::class.java.simpleName
+    override val layoutRes: Int
+        get() = R.layout.fragment_more
 
-/**
- * A simple [Fragment] subclass.
- * Use the [MoreFragment.newInstance] factory method to
- * create an instance of this fragment.
- */
-class MoreFragment : Fragment() {
-    // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
+    private val homeViewModel by activityViewModels<HomeViewModel>()
+    private val requestPermissionLauncher =
+        registerForActivityResult(ActivityResultContracts.RequestPermission()) { isGranted: Boolean ->
+            if (isGranted) {
+                navigateGallery()
+            } else {
+                Toast.makeText(
+                    requireContext(),
+                    "갤러리 접근 권한이 없습니다.",
+                    Toast.LENGTH_SHORT
+                ).show()
+            }
+        }
+    private val galleryLauncher: ActivityResultLauncher<Intent> =
+        registerForActivityResult(ActivityResultContracts.StartActivityForResult()) {
+            if (it.resultCode == RESULT_OK && it.data != null) {
+                val imageUri = it.data?.data
+                homeViewModel.userImage.value = imageUri
+            } else if (it.resultCode == RESULT_CANCELED) {
+                Toast.makeText(requireContext(), "사진 선택이 취소되었습니다", Toast.LENGTH_SHORT).show()
+            }
+        }
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        super.onCreateView(inflater, container, savedInstanceState)
+        binding.viewmodel = homeViewModel
+        binding.lifecycleOwner = viewLifecycleOwner
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        showImageToast()
+        changeProfileImage()
+        clickSave()
+        clickTurnOffAutoLogin()
+        observeEditProfile()
+    }
+
+    private fun showImageToast() {
+        Toast.makeText(requireContext(), "프로필 사진을 변경하려면 사진을 클릭하세요", Toast.LENGTH_SHORT).show()
+    }
+
+    private fun clickSave() {
+        binding.btnSave.setOnClickListener {
+            homeViewModel.editProfile()
         }
     }
 
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_more, container, false)
-    }
-
-    companion object {
-        /**
-         * Use this factory method to create a new instance of
-         * this fragment using the provided parameters.
-         *
-         * @param param1 Parameter 1.
-         * @param param2 Parameter 2.
-         * @return A new instance of fragment MoreFragment.
-         */
-        // TODO: Rename and change types and number of parameters
-        @JvmStatic
-        fun newInstance(param1: String, param2: String) =
-            MoreFragment().apply {
-                arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
+    private fun clickTurnOffAutoLogin() {
+        binding.btnTurnOffAutoLogin.setOnClickListener {
+            homeViewModel.turnOffAutoLogin()
+            homeViewModel.turnOffSuccess.observe(viewLifecycleOwner) {
+                if (it) {
+                    Toast.makeText(requireContext(), "자동로그인 해제", Toast.LENGTH_SHORT).show()
                 }
             }
+        }
+    }
+
+    private fun observeEditProfile() {
+        homeViewModel.updateSuccess.observe(viewLifecycleOwner) {
+            when (it) {
+                true -> {
+                    parentFragmentManager.popBackStack()
+                }
+                else -> {
+                    Toast.makeText(requireContext(), "다시 시도해주세요", Toast.LENGTH_SHORT).show()
+                }
+            }
+        }
+    }
+
+    private fun changeProfileImage() {
+        binding.ivProfileImage.setOnClickListener {
+            when {
+                ContextCompat.checkSelfPermission(
+                    requireContext(),
+                    Manifest.permission.READ_EXTERNAL_STORAGE
+                ) == PackageManager.PERMISSION_GRANTED -> {
+                    navigateGallery()
+                }
+                shouldShowRequestPermissionRationale(Manifest.permission.READ_EXTERNAL_STORAGE) -> {
+                    showInContextUI()
+                }
+                else -> {
+                    requestPermissionLauncher.launch(Manifest.permission.READ_EXTERNAL_STORAGE)
+                }
+            }
+        }
+    }
+
+    private fun navigateGallery() {
+        val intent = Intent(Intent.ACTION_PICK)
+        intent.type = MediaStore.Images.Media.CONTENT_TYPE
+        galleryLauncher.launch(intent)
+    }
+
+    private fun showInContextUI() {
+        AlertDialog.Builder(requireContext())
+            .setTitle("권한 동의 필요")
+            .setMessage("프로필 사진 수정을 위해 갤러리 접근 권한이 필요합니다.")
+            .setPositiveButton("동의") { _, _ ->
+                requestPermissionLauncher.launch(Manifest.permission.READ_EXTERNAL_STORAGE)
+            }
+            .setNegativeButton("거부") { _, _ ->
+                Toast.makeText(requireContext(), "갤러리 접근 권한이 없습니다.", Toast.LENGTH_SHORT).show()
+            }
+            .create()
+            .show()
     }
 }
