+++ app/src/main/res/layout/fragment_my_profile.xml
@@ -4,6 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools">
 
     <data>
+
         <variable
             name="viewmodel"
             type="co.kr.sopt_seminar_30th.presentation.viewmodel.HomeViewModel" />
@@ -16,79 +17,69 @@
 
         <ImageView
             android:id="@+id/iv_profile_image"
+            setProfileImageUri="@{viewmodel.userImage}"
+            android:layout_width="0dp"
+            android:layout_height="0dp"
+            android:contentDescription="@string/description_profile_image"
+            android:scaleType="centerCrop"
             app:layout_constraintBottom_toTopOf="@id/tv_profile_name"
             app:layout_constraintDimensionRatio="1:1"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintWidth_percent="0.25"
-            setProfileImageUri="@{viewmodel.userImage}"
-            android:layout_width="0dp"
-            android:layout_height="0dp"
-            android:contentDescription="@string/description_profile_image"
-            android:scaleType="centerCrop" />
+            app:layout_constraintWidth_percent="0.25" />
 
         <TextView
             android:id="@+id/tv_profile_name"
-            app:layout_constraintBottom_toTopOf="@id/tv_profile_age"
-            app:layout_constraintEnd_toEndOf="@id/iv_profile_image"
-            app:layout_constraintStart_toStartOf="@id/iv_profile_image"
-            app:layout_constraintTop_toBottomOf="@id/iv_profile_image"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginTop="5dp"
             android:text="@{`이름: ` + viewmodel.user.userName}"
             android:textColor="@color/black"
-            android:textSize="16sp" />
+            android:textSize="16sp"
+            app:layout_constraintBottom_toTopOf="@id/tv_profile_age"
+            app:layout_constraintEnd_toEndOf="@id/iv_profile_image"
+            app:layout_constraintStart_toStartOf="@id/iv_profile_image"
+            app:layout_constraintTop_toBottomOf="@id/iv_profile_image" />
 
         <TextView
             android:id="@+id/tv_profile_age"
-            app:layout_constraintBottom_toTopOf="@id/tv_profile_mbti"
-            app:layout_constraintEnd_toEndOf="@id/tv_profile_name"
-            app:layout_constraintStart_toStartOf="@id/tv_profile_name"
-            app:layout_constraintTop_toBottomOf="@id/tv_profile_name"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginTop="5dp"
             android:text="@{viewmodel.user.userAge == 0 ? `나이: 미입력` : `나이: ` + Integer.toString(viewmodel.user.userAge)}"
             android:textColor="@color/black"
-            android:textSize="16sp" />
+            android:textSize="16sp"
+            app:layout_constraintBottom_toTopOf="@id/tv_profile_mbti"
+            app:layout_constraintEnd_toEndOf="@id/tv_profile_name"
+            app:layout_constraintStart_toStartOf="@id/tv_profile_name"
+            app:layout_constraintTop_toBottomOf="@id/tv_profile_name" />
 
         <TextView
             android:id="@+id/tv_profile_mbti"
-            app:layout_constraintBottom_toTopOf="@id/tv_profile_description"
-            app:layout_constraintEnd_toEndOf="@id/tv_profile_name"
-            app:layout_constraintStart_toStartOf="@id/tv_profile_name"
-            app:layout_constraintTop_toBottomOf="@id/tv_profile_age"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginTop="5dp"
             android:text="@{viewmodel.user.userMbti == null ? `MBTI: 미입력` : `MBTI: ` + viewmodel.user.userMbti}"
             android:textColor="@color/black"
-            android:textSize="16sp" />
+            android:textSize="16sp"
+            app:layout_constraintBottom_toTopOf="@id/tv_profile_description"
+            app:layout_constraintEnd_toEndOf="@id/tv_profile_name"
+            app:layout_constraintStart_toStartOf="@id/tv_profile_name"
+            app:layout_constraintTop_toBottomOf="@id/tv_profile_age" />
 
         <TextView
             android:id="@+id/tv_profile_description"
-            app:layout_constraintBottom_toTopOf="@id/btn_edit"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/tv_profile_mbti"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginTop="5dp"
             android:text="@{viewmodel.user.userDescription == null ? `자기소개: 미입력` : `자기소개: ` + viewmodel.user.userDescription}"
             android:textColor="@color/black"
-            android:textSize="16sp" />
-
-        <Button
-            android:id="@+id/btn_edit"
+            android:textSize="16sp"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/tv_profile_description"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:text="@string/text_edit_profile" />
+            app:layout_constraintTop_toBottomOf="@id/tv_profile_mbti" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
