+++ app/src/main/res/layout/fragment_profile.xml
@@ -43,6 +43,7 @@
             android:id="@+id/btn_profile_follower"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
+            android:layout_marginTop="5dp"
             android:layout_marginEnd="5dp"
             android:background="@drawable/selector_profile_button"
             android:text="@string/text_profile_follower"
@@ -55,6 +56,7 @@
             android:id="@+id/btn_profile_repository"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
+            android:layout_marginTop="5dp"
             android:background="@drawable/selector_profile_button"
             android:text="@string/text_profile_repository"
             app:layout_constraintEnd_toEndOf="parent"
