+++ app/src/main/res/values/strings.xml
@@ -27,5 +27,5 @@
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="text_profile">프로필</string>
     <string name="text_home">홈</string>
-    <string name="text_more">카메라</string>
+    <string name="text_more">더보기</string>
 </resources>
